+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/kotlin/multiplatform/targets/HubdleKotlinMultiplatformTvOSExtension.kt
@@ -13,11 +13,6 @@ import javax.inject.Inject
 import org.gradle.api.Action
 import org.gradle.api.Project
 import org.gradle.api.provider.Property
-import org.gradle.kotlin.dsl.configure
-import org.gradle.kotlin.dsl.getValue
-import org.gradle.kotlin.dsl.getting
-import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
-import org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet
 
 @HubdleDslMarker
 public open class HubdleKotlinMultiplatformTvOSExtension @Inject constructor(project: Project) :
@@ -72,39 +67,6 @@ public open class HubdleKotlinMultiplatformTvOSExtension @Inject constructor(pro
                 tvosArm64()
                 tvosSimulatorArm64()
                 tvosX64()
-            }
-        }
-        lazyConfigurable {
-            configure<KotlinMultiplatformExtension> {
-                val commonMain: KotlinSourceSet by sourceSets.getting
-                val tvosArm64Main: KotlinSourceSet? = sourceSets.findByName("tvosArm64Main")
-                val tvosX64Main: KotlinSourceSet? = sourceSets.findByName("tvosX64Main")
-                val tvosSimulatorArm64Main: KotlinSourceSet? =
-                    sourceSets.findByName("tvosSimulatorArm64Main")
-
-                val commonTest: KotlinSourceSet by sourceSets.getting
-                val tvosArm64Test: KotlinSourceSet? = sourceSets.findByName("tvosArm64Test")
-                val tvosX64Test: KotlinSourceSet? = sourceSets.findByName("tvosX64Test")
-                val tvosSimulatorArm64Test: KotlinSourceSet? =
-                    sourceSets.findByName("tvosSimulatorArm64Test")
-
-                val tvosMainSourceSets: List<KotlinSourceSet> =
-                    listOfNotNull(tvosArm64Main, tvosX64Main, tvosSimulatorArm64Main)
-                val tvosTestSourceSets: List<KotlinSourceSet> =
-                    listOfNotNull(tvosArm64Test, tvosX64Test, tvosSimulatorArm64Test)
-
-                val tvosMain = sourceSets.maybeCreate("tvosMain")
-                val tvosTest = sourceSets.maybeCreate("tvosTest")
-
-                tvosMain.dependsOn(commonMain)
-                for (tvosMainSourceSet in tvosMainSourceSets) {
-                    tvosMainSourceSet.dependsOn(tvosMain)
-                }
-
-                tvosTest.dependsOn(commonTest)
-                for (tvosTestSourceSet in tvosTestSourceSets) {
-                    tvosTestSourceSet.dependsOn(tvosTest)
-                }
             }
         }
     }